organize_dataset.py
#!/usr/bin/env python3
"""
Script to organize the COVID-19 dataset from the ieee8023 repository
"""

import os
import pandas as pd
import shutil
from pathlib import Path

def organize_covid_dataset():
    """Organize the downloaded COVID dataset"""
    print("Organizing COVID-19 dataset...")
    
    # Read metadata
    metadata_file = "temp_dataset/metadata.csv"
    if not os.path.exists(metadata_file):
        print("Metadata file not found!")
        return False
    
    df = pd.read_csv(metadata_file)
    print(f"Total records in metadata: {len(df)}")
    
    # Filter for X-ray images only
    xray_df = df[df['modality'] == 'X-ray'].copy()
    print(f"X-ray images: {len(xray_df)}")
    
    # Create data directories
    os.makedirs('data/COVID', exist_ok=True)
    os.makedirs('data/Normal', exist_ok=True)
    
    covid_count = 0
    normal_count = 0
    
    # Process each row
    for idx, row in xray_df.iterrows():
        if pd.isna(row['filename']):
            continue
            
        source_path = f"temp_dataset/images/{row['filename']}"
        
        if not os.path.exists(source_path):
            continue
        
        # Determine if it's COVID or not
        finding = str(row['finding']).lower()
        rt_pcr = str(row['RT_PCR_positive']).upper()
        
        if 'covid' in finding or rt_pcr == 'Y':
            # COVID case
            dest_path = f"data/COVID/covid_{covid_count}_{row['filename']}"
            shutil.copy2(source_path, dest_path)
            covid_count += 1
        elif 'normal' in finding.lower() or 'no finding' in finding.lower():
            # Normal case
            dest_path = f"data/Normal/normal_{normal_count}_{row['filename']}"
            shutil.copy2(source_path, dest_path)
            normal_count += 1
    
    print(f"Organized dataset:")
    print(f"  COVID cases: {covid_count}")
    print(f"  Normal cases: {normal_count}")
    
    # Since we have limited normal cases, let's create some mock normal data for demonstration
    if normal_count < 50:
        print("Creating additional mock normal data for demonstration...")
        import numpy as np
        from PIL import Image
        
        for i in range(100 - normal_count):
            # Create a mock "normal" chest X-ray (random grayscale image)
            mock_image = np.random.randint(0, 255, (256, 256), dtype=np.uint8)
            img = Image.fromarray(mock_image, 'L').convert('RGB')
            img.save(f"data/Normal/mock_normal_{i}.png")
            normal_count += 1
    
    print(f"Final dataset:")
    print(f"  COVID cases: {covid_count}")
    print(f"  Normal cases: {normal_count}")
    
    return True

def download_normal_chest_xrays():
    """Try to download some normal chest X-ray images"""
    print("Attempting to download normal chest X-rays...")
    
    # Try to clone a repository with normal chest X-rays
    try:
        os.system("git clone https://github.com/muhammedtalo/COVID-19 temp_normal --depth 1")
        if os.path.exists("temp_normal"):
            print("Successfully downloaded additional dataset")
            return True
    except:
        pass
    
    return False

if __name__ == "__main__":
    print("COVID-19 Dataset Organization Script")
    print("=" * 40)
    
    # Download additional normal images if possible
    download_normal_chest_xrays()
    
    # Organize the main dataset
    if organize_covid_dataset():
        print("\nDataset organization complete!")
        print("You can now run the Jupyter notebook.")
    else:
        print("\nDataset organization failed.")